+++ NEWS.md
@@ -7,6 +7,7 @@ PENDING: dataiter 0.28
 * `DataFrame.read_csv`: Add argument `dtypes`
 * `DataFrame.read_json`: Add arguments `columns` and `dtypes`
 * `GeoJSON.read`: Add arguments `columns` and `dtypes`
+* `ListOfDicts.fill_missing`: New method
 * `ListOfDicts.from_json`: Add arguments `keys` and `types`
 * `ListOfDicts.full_join`: Speed up
 * `ListOfDicts.read_csv`: Add argument `types`, rename `columns` to `keys`